move capa.features.capabilities to capa.capabilities, and update scripts
