+++ tests/test_capabilities.py
@@ -8,7 +8,7 @@
 # See the License for the specific language governing permissions and limitations under the License.
 import textwrap
 
-import capa.features.capabilities.common
+import capa.capabilities.common
 
 
 def test_match_across_scopes_file_function(z9324d_extractor):
@@ -74,7 +74,7 @@ def test_match_across_scopes_file_function(z9324d_extractor):
             ),
         ]
     )
-    capabilities, meta = capa.features.capabilities.common.find_capabilities(rules, z9324d_extractor)
+    capabilities, meta = capa.capabilities.common.find_capabilities(rules, z9324d_extractor)
     assert "install service" in capabilities
     assert ".text section" in capabilities
     assert ".text section and install service" in capabilities
@@ -142,7 +142,7 @@ def test_match_across_scopes(z9324d_extractor):
             ),
         ]
     )
-    capabilities, meta = capa.features.capabilities.common.find_capabilities(rules, z9324d_extractor)
+    capabilities, meta = capa.capabilities.common.find_capabilities(rules, z9324d_extractor)
     assert "tight loop" in capabilities
     assert "kill thread loop" in capabilities
     assert "kill thread program" in capabilities
@@ -170,7 +170,7 @@ def test_subscope_bb_rules(z9324d_extractor):
         ]
     )
     # tight loop at 0x403685
-    capabilities, meta = capa.features.capabilities.common.find_capabilities(rules, z9324d_extractor)
+    capabilities, meta = capa.capabilities.common.find_capabilities(rules, z9324d_extractor)
     assert "test rule" in capabilities
 
 
@@ -194,7 +194,7 @@ def test_byte_matching(z9324d_extractor):
             )
         ]
     )
-    capabilities, meta = capa.features.capabilities.common.find_capabilities(rules, z9324d_extractor)
+    capabilities, meta = capa.capabilities.common.find_capabilities(rules, z9324d_extractor)
     assert "byte match test" in capabilities
 
 
@@ -219,7 +219,7 @@ def test_count_bb(z9324d_extractor):
             )
         ]
     )
-    capabilities, meta = capa.features.capabilities.common.find_capabilities(rules, z9324d_extractor)
+    capabilities, meta = capa.capabilities.common.find_capabilities(rules, z9324d_extractor)
     assert "count bb" in capabilities
 
 
@@ -246,7 +246,7 @@ def test_instruction_scope(z9324d_extractor):
             )
         ]
     )
-    capabilities, meta = capa.features.capabilities.common.find_capabilities(rules, z9324d_extractor)
+    capabilities, meta = capa.capabilities.common.find_capabilities(rules, z9324d_extractor)
     assert "push 1000" in capabilities
     assert 0x4071A4 in {result[0] for result in capabilities["push 1000"]}
 
@@ -278,6 +278,6 @@ def test_instruction_subscope(z9324d_extractor):
             )
         ]
     )
-    capabilities, meta = capa.features.capabilities.common.find_capabilities(rules, z9324d_extractor)
+    capabilities, meta = capa.capabilities.common.find_capabilities(rules, z9324d_extractor)
     assert "push 1000 on i386" in capabilities
     assert 0x406F60 in {result[0] for result in capabilities["push 1000 on i386"]}
